move test_deploy and test_add_and_remove to their own file
